+++ src/Chart/Markup.hs
@@ -7,6 +7,10 @@ module Chart.Markup
   ( Markup (..),
     ChartOptions (..),
     forgetHud,
+    asChartTree,
+    vertCO,
+    horiCO,
+    stackCO,
     markupChartOptions,
     markupChartTree,
     markupChart,
@@ -211,13 +215,8 @@ attsText o =
           ("fill", showRGB $ view #color o),
           ("fill-opacity", showOpacity $ view #color o),
           ("font-size", encodeNum $ view #size o),
-          ("text-anchor", toTextAnchor $ view #anchor o)
+          ("text-anchor", fromTextAnchor $ view #textAnchor o)
         ]
-  where
-    toTextAnchor :: Anchor -> ByteString
-    toTextAnchor AnchorMiddle = "middle"
-    toTextAnchor AnchorStart = "start"
-    toTextAnchor AnchorEnd = "end"
 
 -- | GlyphStyle to [Attr]
 attsGlyph :: Style -> [Attr]
@@ -261,7 +260,7 @@ markupShape_ SquareGlyph s p =
   markupRect (move p ((s *) <$> one :: Rect Double))
 markupShape_ (RectSharpGlyph x') s p =
   markupRect (move p (scale (Point s (x' * s)) one :: Rect Double))
-markupShape_ (RectRoundedGlyph x' rx ry) s p = emptyElem "rect" as
+markupShape_ (RectRoundedGlyph x' rx' ry') s p = emptyElem "rect" as
   where
     as =
       uncurry Attr
@@ -269,8 +268,8 @@ markupShape_ (RectRoundedGlyph x' rx ry) s p = emptyElem "rect" as
               ("height", encodeNum $ w - y),
               ("x", encodeNum x),
               ("y", encodeNum $ -w),
-              ("rx", encodeNum rx),
-              ("ry", encodeNum ry)
+              ("rx", encodeNum rx'),
+              ("ry", encodeNum ry')
             ]
     (Rect x z y w) = move p (scale (Point s (x' * s)) one)
 markupShape_ (TriangleGlyph (Point xa ya) (Point xb yb) (Point xc yc)) s p =
@@ -483,6 +482,22 @@ forgetHud co =
     & set #hudOptions mempty
     & set #chartTree (addHud (view (#markupOptions % #chartAspect) co) (view #hudOptions co) (view #chartTree co))
     & set (#chartTree % charts' % each % #chartStyle % #scaleP) ScalePArea
+
+-- | Convert ChartOptions to a ChartTree, forgetting the original hud and chart data
+asChartTree :: ChartOptions -> ChartTree
+asChartTree = view #chartTree . forgetHud
+
+-- | Horizontally stack a list of ChartOptions (proceeding to the right), at the supplied Align and with the supplied gap intercalated.
+horiCO :: Align -> Double -> [ChartOptions] -> ChartOptions
+horiCO align gap cs = mempty @ChartOptions & set #chartTree (hori align gap (asChartTree <$> cs))
+
+-- | Vertically stack a list of ChartOptions (proceeding upwards), at the supplied Align and with the supplied gap intercalated.
+vertCO :: Align -> Double -> [ChartOptions] -> ChartOptions
+vertCO align gap cs = mempty @ChartOptions & set #chartTree (vert align gap (asChartTree <$> cs))
+
+-- | Stack a list of ChartOptions horizontally, then vertically (proceeding downwards which is opposite to the usual coordinate reference system but intuitively the way people read charts)
+stackCO :: Int -> Align -> Align -> Double -> [ChartOptions] -> ChartOptions
+stackCO n alignV alignH gap cs = mempty @ChartOptions & set #chartTree (stack n alignV alignH gap (asChartTree <$> cs))
 
 -- | Convert ChartOptions to Markup
 --